@@ -26,51 +26,49 @@
 	},
 	"dependencies": {
 		"@repo/ui": "*",
-		"browser-info": "^1.2.0",
+		"browser-info": "^1.3.0",
 		"clsx": "^2.1.1",
-		"loglevel": "^1.6.1",
-		"query-string": "^6.1.0",
+		"loglevel": "^1.9.2",
+		"query-string": "^9.1.1",
 		"react": "18.3.1",
 		"react-copy-to-clipboard": "^5.0.1",
 		"react-dom": "18.3.1",
-		"react-router": "^4.3.1",
-		"react-router-dom": "^4.3.1",
-		"react-toggle": "^4.0.2"
+		"react-router": "^7.0.1",
+		"react-router-dom": "^7.0.1",
+		"react-toggle": "^4.1.3"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.22.17",
-		"@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.7",
-		"@babel/plugin-proposal-object-rest-spread": "^7.9.6",
-		"@babel/plugin-proposal-optional-chaining": "^7.16.7",
-		"@babel/plugin-syntax-jsx": "^7.2.0",
+		"@babel/core": "^7.26.0",
+		"@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
+		"@babel/plugin-proposal-object-rest-spread": "^7.20.7",
+		"@babel/plugin-proposal-optional-chaining": "^7.21.0",
+		"@babel/plugin-syntax-jsx": "^7.25.9",
 		"@babel/plugin-syntax-optional-chaining": "^7.8.3",
-		"@babel/preset-env": "^7.22.15",
-		"@babel/preset-react": "^7.0.0-beta.49",
+		"@babel/preset-env": "^7.26.0",
+		"@babel/preset-react": "^7.25.9",
 		"@svgr/webpack": "^8.1.0",
 		"@types/react": "^18.3.12",
 		"@types/react-copy-to-clipboard": "^5.0.7",
 		"@types/react-dom": "18.3.1",
-		"@types/react-router": "^4.3.1",
-		"@types/react-router-dom": "^4.3.1",
 		"@types/react-toggle": "^4.0.5",
 		"@types/webextension-polyfill": "^0.12.1",
-		"babel-loader": "^8.0.0-beta.3",
-		"babel-plugin-transform-class-properties": "^6.10.2",
+		"babel-loader": "^9.2.1",
+		"babel-plugin-transform-class-properties": "^6.24.1",
 		"clean-webpack-plugin": "^4.0.0",
 		"copy-webpack-plugin": "^12.0.2",
-		"css-loader": "^5.2.4",
-		"html-webpack-plugin": "^5.5.3",
-		"mini-css-extract-plugin": "^2.2.2",
+		"css-loader": "^7.1.2",
+		"html-webpack-plugin": "^5.6.3",
+		"mini-css-extract-plugin": "^2.9.2",
 		"node-sass": "^9.0.0",
-		"opener": "^1.5.1",
-		"sass-loader": "^14.2.1",
+		"opener": "^1.5.2",
+		"sass-loader": "^16.0.3",
 		"ts-loader": "^9.5.1",
-		"typescript": "^5.6.3",
+		"typescript": "^5.7.2",
 		"webextension-polyfill": "^0.12.0",
-		"webpack": "^5.88.2",
-		"webpack-bundle-analyzer": "^4.9.1",
-		"webpack-cli": "^4.8.0",
-		"webpack-sources": "^3.2.0",
-		"zip-webpack-plugin": "^4.0.1"
+		"webpack": "^5.96.1",
+		"webpack-bundle-analyzer": "^4.10.2",
+		"webpack-cli": "^5.1.4",
+		"webpack-sources": "^3.2.3",
+		"zip-webpack-plugin": "^4.0.2"
 	}
 }
