@@ -13,17 +13,17 @@
     "codegen": "npm run build-codegen && cd ../../packages/codegen && npm start -- ../../apps/frontend/src/codegen/generators ../../apps/frontend/src/codegen/output && cd ../../apps/frontend && prettier --log-level=warn ./src/codegen/output --write"
   },
   "dependencies": {
-    "@hyperjump/json-schema": "^1.8.0",
-    "@next/env": "^14.2.2",
+    "@hyperjump/json-schema": "^1.9.9",
+    "@next/env": "15.0.3",
     "@repo/ui": "*",
-    "ajv": "^8.12.0",
+    "ajv": "^8.17.1",
     "bcp-47": "^2.1.0",
-    "clsx": "^2.1.0",
+    "clsx": "^2.1.1",
     "html5-qrcode": "^2.3.8",
-    "iso-639-1": "^3.1.2",
-    "nanoid": "^5.0.7",
-    "negotiator": "^0.6.3",
-    "next": "^14.2.2",
+    "iso-639-1": "^3.1.3",
+    "nanoid": "^5.0.8",
+    "negotiator": "^1.0.0",
+    "next": "15.0.3",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "server-only": "^0.0.1",
@@ -31,15 +31,15 @@
   },
   "devDependencies": {
     "@types/negotiator": "^0.6.3",
-    "@types/node": "^20.12.7",
+    "@types/node": "^20.9.0",
     "@types/react": "^18.3.12",
     "@types/stringify-object": "^4.0.5",
-    "eslint": "8.56.0",
-    "eslint-config-next": "14.2.2",
+    "eslint": "9.15.0",
+    "eslint-config-next": "15.0.3",
     "eslint-config-prettier": "^9.1.0",
-    "prettier": "3.2.5",
-    "sass": "^1.75.0",
+    "prettier": "3.3.3",
+    "sass": "^1.81.0",
     "serve": "^14.2.4",
-    "typescript": "^5.4.5"
+    "typescript": "^5.7.2"
   }
 }
