@@ -14,8 +14,8 @@
     "#lib/*": "./dist/src/lib/*/index.js"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.7.0",
+    "@biomejs/biome": "1.9.4",
     "@types/node": "^20.12.7",
-    "typescript": "^5.4.5"
+    "typescript": "^5.7.2"
   }
 }
