@@ -1,5 +1,5 @@
 import React from "react";
-import { Route, Switch } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 // @ts-expect-error no type info for this
 import browserInfo from "browser-info";
 import SettingsPage from "./SettingsPage";
@@ -8,15 +8,18 @@ import InformationPage from "./InformationPage";
 
 import "./ContentsArea.scss";
 
-const isValidShortcuts = browserInfo().name == "Firefox" && browserInfo().version >= 60;
+const isValidShortcuts =
+  browserInfo().name == "Firefox" && browserInfo().version >= 60;
 
 export default () => (
   <div className="contentsArea">
-    <Switch>
-      <Route path="/settings" component={SettingsPage} />
-      {isValidShortcuts && <Route path="/shortcuts" component={KeyboardShortcutsPage} />}
-      <Route path="/information" component={InformationPage} />
-      <Route component={SettingsPage} />
-    </Switch>
+    <Routes>
+      <Route path="/settings" Component={SettingsPage} />
+      {isValidShortcuts && (
+        <Route path="/shortcuts" Component={KeyboardShortcutsPage} />
+      )}
+      <Route path="/information" Component={InformationPage} />
+      <Route Component={SettingsPage} />
+    </Routes>
   </div>
 );
